proyectoNCapas

+++ src/main/java/com/digis/equipo2/Equipo2ControlEscolar/JPARepository/AlumnoRepository.java
@@ -12,6 +12,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
  *
  * @author digis
  */
-    public interface AlumnoRepository extends JpaRepository <Alumno, Long> {
+    public interface AlumnoRepository extends JpaRepository <Alumno, Integer> {
     
 }
